wip: simple envs feature
