+++ packages/polywrap-client-config-builder/tests/test_client_config_builder.py
@@ -1,4 +1,3 @@
-from multiprocessing.connection import Client
 from typing import List
 from polywrap_core.types.env import Env
 from polywrap_core.types.uri import Uri
@@ -11,26 +10,29 @@ test_envs: List[Env] = [
     Env(uri = Uri("wrap://ens/test.plugin.two"), env = { 'test': "value" }),
   ]
 
-client = ClientConfigBuilder()
 
-def test_client_config_builder_adds_envs(test_envs: List[Env]):
+def test_client_config_builder_add_env():
+    client = ClientConfigBuilder().add_env(
+        Uri("wrap://ens/test.plugin.one"),
+        Env(uri = Uri("wrap://ens/test.plugin.one"), env = { 'test': "value" }),
+        )
+    print(client._config['envs'])
+
     print(client)
     for env in test_envs:
         print('adding an env', env)
-        client.add_env(env.uri, env.env)
+        client = client.add_env(env.uri, env.env)
     
     print(client)
-    return client
+    return False
 
-def test_client_config_builder_adds_default_config():
-    print(client.build_partial())
-    # print(client.add_defaults().build_partial())
-    print(type(client.build_partial()))
-    print(client.envs)
-    print(type(client.build()))
-
-    assert False
-
+# def test_client_config_builder_adds_default_config():
+#     #print(client.build_partial())
+#     # print(client.add_defaults().build_partial())
+#     #print(type(client.build_partial()))
+#     print(client.envs)
+#     print(type(client.build()))
+#     pass
 # def test_client_config_builder_adds_config():
 #     envs: List[Env] = []
 #     resolver: IUriResolver = IUriResolver()
